Handle flush events and fix iterator
